[pending rendering test] move text-layer to core-layers
