+++ src/core-layers/text-layer/multi-icon-layer/multi-icon-layer-vertex.glsl.js
