+++ src/experimental-layers/src/index.js
@@ -9,7 +9,6 @@ export {default as PathOutlineLayer} from './path-outline-layer/path-outline-lay
 export {default as SolidPolygonLayer} from './solid-polygon-layer/solid-polygon-layer';
 
 export {default as Arrow2DGeometry} from './path-marker-layer/arrow-2d-geometry';
-export {default as TextLayer} from './text-layer/text-layer';
 export {default as AdvancedTextLayer} from './advanced-text-layer/advanced-text-layer';
 
 export {default as outline} from './shaderlib/outline/outline';
